Moved @bangle.io and @bangle.dev packages inside components/editor
