+++ components/editor/@bangle.io/extensions/inline-command-palette/config.ts
