+++ components/editor/@bangle.io/extensions/inline-command-palette/palette-item.ts
