+++ components/editor/@bangle.io/js-lib/inline-palette/index.ts
