+++ components/editor/@bangle.io/js-lib/inline-palette/inline-palette.ts
