+++ components/editor/@bangle.io/lib/config/index.ts
