+++ components/editor/@bangle.io/lib/config/package.json
