+++ components/editor/@bangle.io/lib/ui-components/UniversalPalette/index.ts
